Relocate `DomainClassArtefactHandler` to grace-plugin-domain-class
