+++ grace-plugin-converters/build.gradle
@@ -21,6 +21,7 @@ dependencies {
     annotationProcessor libs.spring.boot.autoconfigureProcessor
     annotationProcessor libs.spring.boot.configurationProcessor
 
+    testImplementation project(":grace-plugin-domain-class")
     testImplementation libs.grace.datastore.gorm.support
     testImplementation libs.jakarta.servlet
     testImplementation libs.json.unit
