+++ grace-plugin-domain-class/src/main/groovy/grails/core/GrailsDomainClass.java
