+++ grace-plugin-domain-class/src/main/groovy/org/grails/core/artefact/DomainClassArtefactHandler.java
