+++ grace-plugin-validation/build.gradle
@@ -3,6 +3,7 @@ dependencies {
     api project(":grace-util")
     api libs.grace.datastore.gorm.validation
 
+    testImplementation project(":grace-plugin-domain-class")
     testImplementation project(":grace-web-common")
     testImplementation libs.jakarta.servlet
     testImplementation libs.spring.test
