+++ grace-web-gsp/build.gradle
@@ -2,6 +2,7 @@ dependencies {
     compileOnly libs.ant
     compileOnlyApi libs.jakarta.servlet
     api project(":grace-gsp")
+    compileOnly project(":grace-plugin-domain-class")
     api project(":grace-web-common")
     api project(":grace-web-taglib")
     api project(":grace-web-sitemesh")
